Choose an arbitrary second in the range [0; 60]
It's an older change. As far as I remember, it's intended to help with
occasionally failing tests, although this may not be enough because the
expected precision of a second in `GFS` is one second, i.e. we don't
care about the fractional part.

The range [0; 60] is due to this comment in `TimeOfDay`:

> Note that 0 <= `todSec` < 61, accomodating leap seconds.