Move more common code into `arbitraryBaseTestData`
`BaseTestData` isn't parameterized anymore because I'm not sure it's
necessary and also this way, it's easier; although it knows about
`NumSubperiods` and `NewestTimes` even if a specific test function
doesn't need those values (and returning an empty list for `NewestTimes`
is a hack).

Now the `arbitraryInput*` generators only generate the number of
subperiods and arbitrary offsets, so a lot of duplicate code is gone.
However `NewestTimes` is always a `Float` now, which moved a few
`ceiling` calls downstream (an alternative is to hide the type behind a
parameter that could be `Int` or `Float`, but that didn't work in an
easy way).

I guess I like the results so far, although I'm still not sure if this
refactored generator code will be easier to extend (for multiple
periods).