Verify the property of max removed times after a shift in `now`
This property is an important one to confirm (just for my own sanity),
i.e. if you run the program with the same settings a day (depends on the
settings) later, it won't remove more than 2 items.

Once again, QuickCheck to the rescue here! I initially thought of the
max 1 time removed in the test, but QuickCheck quickly found a
counterexample with an incomplete subperiod (see the code comment).