Verity the max removed items after a `now` shift given equal subperiods
This follows the previous test and confirms the same property when
number of subperiods is an integer! This is very easy because the same
`arbitraryInputWithinRangeSubperiods` can be used to generate data for
both cases! (Although the test code is still duplicated).

A simple `Num a` constraint didn't work yesterday, and adding a
`RealFrac a` didn't work for `Int`, so the solution is create a custom
typeclass that can represent the number of subperiods and can be
implemented by `Int` and `Float`. It needs to define the required
polymorphic functions (this is very close to the Interface Segregation
Principle).