Support cleaning up in multiple periods
It's not fully implemented yet because it expects that there is at least
one time in every period (`zip periods groupedByPeriods`).

The `cleanup_` version with the log output is also exported now — it's
helpful to print the log in case a test fails (implemented only in the
new test). It may also be useful in the future to users running the
program in verbose mode.