[wip] Remove `f` type parameter from `BaseTestData`
There is no need in it anymore because all cases used `Identity`.

Previously, when `cleanup` accepted a list of periods, this `f`
parameter told the testing code (`GFSSpec`) whether there was only one
period (wrapped in `Identity`) or multiple periods (wrapped in `[]`).
Now `PeriodInfo` includes `Offsets`, so it's always a non-empty list of
offsets.