Fix including `newestOffset` in `arbitraryOffsets`
The function comment stated "newest offsets are also included in all
offsets", which was its intention, but in fact it did not and the users
of the function manually added `newestOffsets` to `offsets`. Now that
responsibility is moved into `arbitraryOffsets` as it should be — the
function knows the relative positions of the offsets, so sorting them
there should be more efficient than sorting the whole bunch of offsets
at the end.