Don't produce checkpoints older than the oldest time from biggest limit
This is an interesting failure discovered with seed 1604951951, where
subtracting a bigger interval produced a newer time:

2023-01-02 14:29:00 - 70m70h = 2017-02-28 09:29:00
2023-01-02 14:29:00 - 70m91h = 2017-02-28 19:29:00

It shouldn't happen with the default ranges, so it wasn't clear how to
handle this. The fix here is to make sure all produced checkpoints are
not older than the should-be-oldest checkpoint calculated from the
biggest interval limit.