Fix cleaning up "day 31" backups
In fact, there is nothing special about the day 31, I've only noticed
this bug for those days on a few occasions. The bug is that the newest
of the backups decided to be removed was kept, whereas it should be the
newest backup in the input list, before now. `keepNewestTime` was placed
too late in the processing; on the other hand, it can't be put on the
input list either because there could be future dates there, so the
right place is in `gfsRemove` when we know the backups that will be
filtered (that is, before now).